Enhance divergence detection with improved analysis and reporting
