+++ diff_analysis/diff_events.csv
@@ -1,8 +1,3 @@
-step,signal,z_score,run_a_value,run_b_value,violation_type
-29,kl_mean,-1.5624769985660254,0.1258,0.137,lower
-97,kl_mean,-2.1846607310643766,0.1912,0.6194999999999999,lower
-100,kl_mean,-1.9045569462576513,0.2072,0.5907,lower
-103,kl_mean,-1.8117252355419355,0.2116,0.5801999999999999,lower
-134,kl_mean,-1.2441940571143755,0.2305,0.2398,upper
-150,kl_mean,0.9195702294329283,0.2595,0.2521,upper
-156,kl_mean,1.696352647614194,0.262,0.2475,lower
+step,signal,z_score,run_a_value,run_b_value,violation_type,consecutive_count
+103,kl_mean,-2.033059304661903,0.1974,0.5940000000000001,lower,3
+45,reward_mean,2.0969826179560562,0.9905,0.9057,upper,3
